treeProject
I made a file: treeProject to use as our code to show. My files are all messed up from all the times we had to edit classes, so I basically just took code from the internet to create the inorder method. it just needs to be modified to match our library